Split up code base on commonmark.js
